Rename `df_creator.py` to `transformations.py`
Not only is this more descriptive of the module's function, it is also better directory structure, as the `data` sub-directory is renamed to `dataframe`
